% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Airportwise Airtime.R
\name{Avg_time_airport}
\alias{Avg_time_airport}
\title{Calculate the average airtime of all the filghts airport-wise}
\usage{
Avg_time_airport(origin, Destination, month = c(1:12))
}
\arguments{
\item{Destination}{Destination Airport}

\item{Origin}{Origin Airport}

\item{Month}{Month of flight}
}
\value{
Average of the total airtime for the flight data in hours
}
\description{
Calculate the average airtime of all the filghts airport-wise
}
\examples{
Avg_time_airport("JFK", "LAX")
}

