% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statewise Airtime.R
\name{Avg_time_state}
\alias{Avg_time_state}
\title{Calculate the average airtime of all the filghts state-wise}
\usage{
Avg_time_state(origin, Destination, months = c(1:12))
}
\arguments{
\item{Destination}{Destination Airport}

\item{Origin}{Origin Airport}

\item{Months}{Month of Departure}
}
\value{
Average of the total airtime for the flight data
}
\description{
Calculate the average airtime of all the filghts state-wise
}
\examples{
Avg_time_state("CA", "CO")

}

