% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delay_Airport.R
\name{Delay_Airport}
\alias{Delay_Airport}
\title{Calculate the delay time by airport}
\usage{
Delay_Airport(origin, destination = NULL, month = c(1:12))
}
\arguments{
\item{Origin}{State they are leaving from.}

\item{Destination}{State they are going to.}

\item{Month}{Month of Departure}
}
\value{
Estimated delay time based one the state the user is going in minutes
}
\description{
Calculate the delay time by airport
}
\examples{
Delay_Airport("NY", "MI")
Delay_Airport("CO")
}

