% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delay_Reason_Airport.R
\name{Delay_Reason}
\alias{Delay_Reason}
\title{Returns the most common cause for delay in flights between the given airports}
\usage{
Delay_Reason(Origin, Destination = NULL)
}
\arguments{
\item{Origin}{Origin airport abbreviation}

\item{Destination}{Destination airport abbreviation}
}
\value{
The most frequent cause of delay
}
\description{
Returns the most common cause for delay in flights between the given airports
}
\examples{
Delay_Reason("JFK")
Delay_Reason("JFK", "LAX")
}

