% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delay_Reason_State.R
\name{Delay_Reason_State}
\alias{Delay_Reason_State}
\title{Calculate the area of an ellipse.}
\usage{
Delay_Reason_State(Origin, Destination = NULL)
}
\arguments{
\item{Origin}{Origin state abbreviation}

\item{Destination}{Destination state abbreviation}
}
\value{
The most common cause for delay in flights between the given states
}
\description{
Calculate the area of an ellipse.
}
\examples{
Delay_Reason_State("NJ")
Delay_Reason_State("NJ", "CO")
}

